rearranged code; nav now increments until 255 and stops
